feat: deprecate destination (in favour of source) (#8)
The original codebase corresponds to a Vesting Contract between A and B,
in which A gives tokens to B in a scheduled way. The intended behaviour
of this proposal is for A to lock tokens, that only A can withdraw.

Following that intended behaviour, we deprecate the concept of
"destination" from every call (as it will be same as the source).